Final commit
Finished project. All functions reside in player, gameboard, or display
umbrella IIFE's. This is easily the most complex and functional code
that I've written to date. I must say that the modular code is so much
cleaner and useful once you get the hang of it. Immensely proud of this
project, may go back someday and add a scoreboard, but for now I am
eager to learn more concepts. I know exactly how I would implement the
scoreboard if I wanted to though, which is exciting.